バグの修正
残りの課題
割り算の際のオーバーフロー

+++ ge_p6/src/stack_display.c
@@ -11,6 +11,6 @@
 void stack_display(struct stack *five){
 	int i;
 	for(i=0;i<five->head;i++){
-		printf("stack[%d]:%d\n",i,(int)five->value[i]);
+		printf("stack[%d]:%lld\n",i,five->value[i]);
 	}
 }
